Reworked ConfigReader to accept non-flyway entries in db.*
Previous ConfigReader would throw an exception when users attempted to configure the Play pool, or provided any setting under the db group that did not contain both a URL and a Driver specification.

New ConfigReader ignores settings that are not groups, and groups that do not contain either a URL or a Driver, and issues warnings only on those that contain a URL or Driver, but not the other half of the necessary pair.

Added a new setting to force flyway to ignore a setting group:
db.*.migration.flywayIgnore = true